REFACTOR: split server.py and deps.py
