+++ src/endfield_essence_recognizer/__init__.py
@@ -1,3 +1,9 @@
 from __future__ import annotations
 
+from endfield_essence_recognizer.server import app
 from endfield_essence_recognizer.version import __version__ as __version__
+
+__all__ = [
+    "__version__",
+    "app",
+]
