+++ src/endfield_essence_recognizer/api/router.py
@@ -1,0 +1,13 @@
+from fastapi import APIRouter
+
+from .routes import config, scanner, screenshot, system
+from .websockets import logs
+
+api_router = APIRouter(prefix="/api")
+api_router.include_router(config.router)
+api_router.include_router(scanner.router)
+api_router.include_router(screenshot.router)
+api_router.include_router(system.router)
+
+ws_router = APIRouter(prefix="/ws")
+ws_router.include_router(logs.router)
