--- src/endfield_essence_recognizer/deps.py
@@ -1,352 +1,0 @@
-"""
-FastAPI dependency injection.
-"""
-
-from functools import lru_cache
-from pathlib import Path
-
-from fastapi import Depends
-
-from endfield_essence_recognizer.core.delivery_claimer.engine import (
-    DeliveryClaimerEngine,
-)
-from endfield_essence_recognizer.core.layout.base import ResolutionProfile
-from endfield_essence_recognizer.core.layout.factory import (
-    build_resolution_profile_strict,
-)
-from endfield_essence_recognizer.core.path import get_config_path, get_screenshots_dir
-from endfield_essence_recognizer.core.recognition import (
-    AbandonStatusRecognizer,
-    AttributeLevelRecognizer,
-    AttributeRecognizer,
-    DeliveryJobRewardRecognizer,
-    DeliverySceneRecognizer,
-    LockStatusRecognizer,
-    UISceneRecognizer,
-    prepare_abandon_status_recognizer,
-    prepare_attribute_level_recognizer,
-    prepare_attribute_recognizer,
-    prepare_delivery_job_reward_recognizer,
-    prepare_delivery_scene_recognizer,
-    prepare_lock_status_recognizer,
-    prepare_ui_scene_recognizer,
-)
-from endfield_essence_recognizer.core.scanner.context import (
-    ScannerContext,
-)
-from endfield_essence_recognizer.core.scanner.engine import ScannerEngine
-from endfield_essence_recognizer.core.webui import get_webview_title
-from endfield_essence_recognizer.core.window import (
-    SUPPORTED_WINDOW_TITLES,
-    WindowManager,
-)
-from endfield_essence_recognizer.core.window.adapter import WindowActionsAdapter
-from endfield_essence_recognizer.exceptions import (
-    UnsupportedResolutionError,
-)
-from endfield_essence_recognizer.services.audio_service import (
-    AudioService,
-    build_audio_service_profile,
-)
-from endfield_essence_recognizer.services.log_service import LogService
-from endfield_essence_recognizer.services.scanner_service import ScannerService
-from endfield_essence_recognizer.services.screenshot_service import ScreenshotService
-from endfield_essence_recognizer.services.user_setting_manager import UserSettingManager
-
-# AudioService dependency
-
-
-@lru_cache
-def get_audio_service() -> AudioService:
-    """
-    Get the AudioService singleton.
-    """
-    return AudioService(build_audio_service_profile())
-
-
-# WindowManager dependency
-
-
-@lru_cache
-def get_window_manager_singleton() -> WindowManager:
-    """
-    Get the singleton WindowManager instance.
-    """
-    return WindowManager(SUPPORTED_WINDOW_TITLES)
-
-
-def get_window_manager_dep() -> WindowManager:
-    """
-    Get the WindowManager dependency.
-    """
-    return get_window_manager_singleton()
-
-
-def get_resolution_profile_dep(
-    window_manager: WindowManager = Depends(get_window_manager_dep),
-) -> ResolutionProfile:
-    """
-    FastAPI dependency: The layout configuration corresponding to the current game window resolution.
-
-    Raises:
-        WindowNotFoundError: If the target window is not found.
-        ValueError: If the screen width or height is not a positive integer.
-        UnsupportedResolutionError: If the screen resolution is unsupported.
-    """
-    width, height = window_manager.get_client_size()  # may raise WindowNotFoundError
-    match build_resolution_profile_strict(width, height):  # may raise ValueError
-        case None:
-            # unsupported resolution
-            raise UnsupportedResolutionError(
-                f"Unsupported resolution: {width}x{height}"
-            )
-        case profile:
-            return profile
-
-
-def get_resolution_profile() -> ResolutionProfile:
-    """
-    A non-dependency version of get_resolution_profile_dep.
-    """
-    return get_resolution_profile_dep(window_manager=get_window_manager_singleton())
-
-
-@lru_cache
-def get_webview_window_manager() -> WindowManager:
-    """
-    Get the singleton WindowManager instance for the webview window.
-    """
-    # Only contain the single webview title
-    return WindowManager([get_webview_title()])
-
-
-# Path dependency
-
-
-def get_config_path_dep() -> Path:
-    """
-    The dependency to get the config path.
-    """
-    return get_config_path()
-
-
-def get_screenshots_dir_dep() -> Path:
-    """
-    The dependency to get the screenshots directory path.
-    """
-    return get_screenshots_dir()
-
-
-# UserSettingManager dependency
-
-
-@lru_cache
-def _get_user_setting_manager_cached(file: Path) -> UserSettingManager:
-    return UserSettingManager(user_setting_file=file)
-
-
-def get_user_setting_manager_singleton(user_setting_file: Path) -> UserSettingManager:
-    """
-    Get the singleton UserSettingManager instance.
-
-    The lru_cache enables testing with different paths in parallel.
-    """
-    absolute_path = user_setting_file.resolve()
-    # Avoid different path representations of the same file
-    return _get_user_setting_manager_cached(absolute_path)
-
-
-def get_user_setting_manager_dep(
-    user_setting_file: Path = Depends(get_config_path_dep),
-) -> UserSettingManager:
-    """
-    Get the singleton UserSettingManager instance.
-    """
-    return get_user_setting_manager_singleton(user_setting_file)
-
-
-def default_user_setting_manager() -> UserSettingManager:
-    """
-    Get the default singleton UserSettingManager instance.
-    """
-    return get_user_setting_manager_singleton(get_config_path())
-
-
-# Recognizer dependencies
-
-
-def get_attribute_recognizer_dep() -> AttributeRecognizer:
-    """
-    Get the default attribute Recognizer instance.
-    """
-    return prepare_attribute_recognizer()
-
-
-def get_attribute_level_recognizer_dep() -> AttributeLevelRecognizer:
-    """
-    Get the default attribute level Recognizer instance.
-    """
-    return prepare_attribute_level_recognizer()
-
-
-def get_abandon_status_recognizer_dep() -> AbandonStatusRecognizer:
-    """
-    Get the default abandon status Recognizer instance.
-    """
-    return prepare_abandon_status_recognizer()
-
-
-def get_lock_status_recognizer_dep() -> LockStatusRecognizer:
-    """
-    Get the default lock status Recognizer instance.
-    """
-    return prepare_lock_status_recognizer()
-
-
-def get_ui_scene_recognizer_dep() -> UISceneRecognizer:
-    """
-    Get the default UI scene Recognizer instance.
-    """
-    return prepare_ui_scene_recognizer()
-
-
-def get_delivery_scene_recognizer_dep() -> DeliverySceneRecognizer:
-    """
-    Get the default delivery scene Recognizer instance.
-    """
-    return prepare_delivery_scene_recognizer()
-
-
-def get_delivery_job_reward_recognizer_dep() -> DeliveryJobRewardRecognizer:
-    """
-    Get the default delivery job reward Recognizer instance.
-    """
-    return prepare_delivery_job_reward_recognizer()
-
-
-# Scanner-related dependencies
-
-
-def default_scanner_context() -> ScannerContext:
-    """
-    Get the default ScannerContext instance.
-
-    Some functions need a ScannerContext but are not called within FastAPI request context.
-    So we provide this default builder function.
-    """
-    return ScannerContext(
-        attr_recognizer=prepare_attribute_recognizer(),
-        attr_level_recognizer=prepare_attribute_level_recognizer(),
-        abandon_status_recognizer=prepare_abandon_status_recognizer(),
-        lock_status_recognizer=prepare_lock_status_recognizer(),
-        ui_scene_recognizer=prepare_ui_scene_recognizer(),
-    )
-
-
-def get_scanner_context_dep(
-    attr_recognizer: AttributeRecognizer = Depends(get_attribute_recognizer_dep),
-    attr_level_recognizer: AttributeLevelRecognizer = Depends(
-        get_attribute_level_recognizer_dep
-    ),
-    abandon_status_recognizer: AbandonStatusRecognizer = Depends(
-        get_abandon_status_recognizer_dep
-    ),
-    lock_status_recognizer: LockStatusRecognizer = Depends(
-        get_lock_status_recognizer_dep
-    ),
-    ui_scene_recognizer: UISceneRecognizer = Depends(get_ui_scene_recognizer_dep),
-) -> ScannerContext:
-    """
-    Get a ScannerContext instance.
-    """
-    return ScannerContext(
-        attr_recognizer=attr_recognizer,
-        attr_level_recognizer=attr_level_recognizer,
-        abandon_status_recognizer=abandon_status_recognizer,
-        lock_status_recognizer=lock_status_recognizer,
-        ui_scene_recognizer=ui_scene_recognizer,
-    )
-
-
-def get_scanner_engine_dep(
-    ctx: ScannerContext = Depends(get_scanner_context_dep),
-    window_manager: WindowManager = Depends(get_window_manager_dep),
-    user_setting_manager: UserSettingManager = Depends(get_user_setting_manager_dep),
-    profile: ResolutionProfile = Depends(get_resolution_profile_dep),
-) -> ScannerEngine:
-    """
-    Get a ScannerEngine instance.
-
-    As this depends on the user setting manager and the profile, it will automatically
-    get the newest UserSettingManager and a ResolutionProfile that matches the current window.
-
-    Raises:
-        WindowNotFoundError: If the target window is not found (from get_resolution_profile_dep).
-        UnsupportedResolutionError: If the screen resolution is unsupported
-            (from get_resolution_profile_dep).
-        ValueError: If the screen width or height is not a positive integer
-    """
-    adapter = WindowActionsAdapter(window_manager)
-    return ScannerEngine(
-        ctx=ctx,
-        image_source=adapter,
-        window_actions=adapter,
-        user_setting_manager=user_setting_manager,
-        profile=profile,
-    )
-
-
-def default_scanner_engine() -> ScannerEngine:
-    """
-    Get the default ScannerEngine instance.
-
-    Some functions need a ScannerEngine but are not called within FastAPI request context.
-
-    Raises:
-        WindowNotFoundError: If the target window is not found.
-        UnsupportedResolutionError: If the screen resolution is unsupported.
-    """
-    window_manager = get_window_manager_singleton()
-    adapter = WindowActionsAdapter(window_manager)
-    profile = get_resolution_profile()  # may raise exceptions
-    return ScannerEngine(
-        ctx=default_scanner_context(),
-        image_source=adapter,
-        window_actions=adapter,
-        user_setting_manager=default_user_setting_manager(),
-        profile=profile,
-    )
-
-
-def default_delivery_claimer_engine() -> DeliveryClaimerEngine:
-    """
-    Get the default DeliveryClaimerEngine instance.
-    """
-    window_manager = get_window_manager_singleton()
-    adapter = WindowActionsAdapter(window_manager)
-    return DeliveryClaimerEngine(
-        image_source=adapter,
-        window_actions=adapter,
-        profile=get_resolution_profile(),  # may raise exceptions
-        delivery_scene_recognizer=get_delivery_scene_recognizer_dep(),
-        delivery_job_reward_recognizer=get_delivery_job_reward_recognizer_dep(),
-        audio_service=get_audio_service(),
-    )
-
-
-@lru_cache
-def get_scanner_service() -> ScannerService:
-    return ScannerService()
-
-
-@lru_cache
-def get_log_service() -> LogService:
-    return LogService()
-
-
-@lru_cache
-def get_screenshot_service() -> ScreenshotService:
-    """
-    Get the ScreenshotService singleton.
-    """
-    return ScreenshotService(get_window_manager_singleton())
